added image parser

+++ obj_detection/parse_pipeline_config.py
@@ -30,41 +30,44 @@ def pipeline_config_to_dict(path_to_file):
 
 def adjust_pipeline_config(path_to_file, overrides_dict):
   print('Adjusting model pipeline.config')
-  try:
-    with open(path_to_file, mode='rt') as f:
-      config = f.read()
+  if os.path.isfile(path_to_file):
+    try:
+      with open(path_to_file, mode='rt') as f:
+        config = f.read()
 
-    for param in overrides_dict.keys():
-      value = overrides_dict[param]
+      for param in overrides_dict.keys():
+        value = overrides_dict[param]
 
-      if type(value) == dict:
-        subDictName, subDict = param, value
+        if type(value) == dict:
+          subDictName, subDict = param, value
 
-        for param in subDict.keys():
-          value = subDict[param]
-          regex = re.escape(subDictName) + r'[\s\w\n{}:"/.]*?' + re.escape(param) + r'(:[ "\w/.:]+)'
-          regex_result = re.search(regex, config)
-          sub_config = regex_result.group(0)
+          for param in subDict.keys():
+            value = subDict[param]
+            regex = re.escape(subDictName) + r'[\s\w\n{}:"/.]*?' + re.escape(param) + r'(:[ "\w/.:]+)'
+            regex_result = re.search(regex, config)
+            sub_config = regex_result.group(0)
+
+            if type(value) == str:
+              value = value.replace('\\','/')
+              new_sub_config = sub_config.replace(regex_result.groups()[0], ': "{}"'.format(value))
+            else:
+              new_sub_config = sub_config.replace(regex_result.groups()[0], ': {}'.format(value))
+
+            config = config.replace(sub_config, new_sub_config)
+        else:
+          regex = re.escape(param) + r': [0-9\w"/.:]+'
 
           if type(value) == str:
-            value = value.replace('\\','/')
-            new_sub_config = sub_config.replace(regex_result.groups()[0], ': "{}"'.format(value))
+            config = re.sub(regex, '{}: "{}"'.format(param, value.replace('\\','/')), config)
           else:
-            new_sub_config = sub_config.replace(regex_result.groups()[0], ': {}'.format(value))
+            config = re.sub(regex, '{}: {}'.format(param, value), config)
 
-          config = config.replace(sub_config, new_sub_config)
-      else:
-        regex = re.escape(param) + r': [0-9\w"/.:]+'
+      with open(path_to_file, mode='wt') as f:
+        f.write(config)
 
-        if type(value) == str:
-          config = re.sub(regex, '{}: "{}"'.format(param, value.replace('\\','/')), config)
-        else:
-          config = re.sub(regex, '{}: {}'.format(param, value), config)
+      print('Successfully adjusted pipeline config')
 
-    with open(path_to_file, mode='wt') as f:
-      f.write(config)
-
-    print('Successfully adjusted pipeline config')
-
-  except Exception as e:
-    print('Failed to adjust pipeline config: ' + str(e))
+    except Exception as e:
+      print('Failed to adjust pipeline config: ' + str(e))
+  else:
+    print('Error: File not found: {}'.format(path_to_file) )
